display comment author with name from ssb-about
