+++ index.html
@@ -58,7 +58,7 @@
         <div id="notifications"></div>
         <div id="walletInfo">
           <b><span id="walletName"></span></b> -
-          <span id="requiredCosigners"></span> of <span id="numberCosigners"></span> multisig. -
+          <span id="requiredCosigners"></span> of <span id="numberCosigners"></span> signatures required.
           Confirmed balance: <span class="balance"></span> BTC
         </div>
         <div id="tabs">
